refactor(CodeEditor): reorganize extensions and update imports for consistency
