+++ packages/code-editor/src/CodeEditor/hooks/extensions/useFoldGutterExtension.tsx
