+++ packages/code-editor/src/CodeEditor/hooks/index.ts
@@ -1,0 +1,3 @@
+export * from './extensions';
+export { useLazyModules } from './useLazyModules';
+export { useModuleLoaders } from './useModuleLoaders';
